break out potential-write-proxy.ts from metamodel.ts (using an agent)
